Move osinfo under internal utils
